requirements.txt
flask
tensorflow==2.18.0
keras==3.5.0
pillow
matplotlib
numpy<2.0
opencv-python
h5py==3.11.0
optree==0.17.0
